sia-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.34
aiomysql==0.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
greenlet>=3.0.3
pydantic==2.11.5
pydantic[email]==2.11.5